requirements.txt
aniso8601==10.0.1
attrs==25.3.0
blinker==1.9.0
click==8.2.1
Flask==2.3.3
flask-restx==1.3.0
gunicorn==21.2.0
importlib_resources==6.5.2
itsdangerous==2.2.0
Jinja2==3.1.6
joblib==1.3.2
jsonschema==4.24.0
jsonschema-specifications==2025.4.1
MarkupSafe==3.0.2
numpy==1.26.4
packaging==25.0
pandas==2.2.3
python-dateutil==2.9.0.post0
python-dotenv==1.0.0
pytz==2025.2
referencing==0.36.2
rpds-py==0.25.1
scikit-learn==1.7.0
scipy==1.16.0
setuptools==80.9.0
six==1.17.0
threadpoolctl==3.6.0
typing_extensions==4.14.0
tzdata==2025.2
Werkzeug==2.3.7
wheel==0.45.1

Django>=4.2.0
djangorestframework>=3.14.0
django-cors-headers>=4.0.0
python-decouple>=3.8